requirements.txt
pillow==10.2.0
google-generativeai==0.4.0
uvicorn==0.27.1
fastapi==0.110.0
python-multipart==0.0.9